First test of using git submodule for the sources
Various experiments with earlier forms of this repository indicated that
as it grows, the performance of git will begin to get unwieldy as we
include more and more large source trees in one repo.  Instead, let's
see if modules can work for this particular use case.

However, rather than point directly at 3rd party repositories, we're
going to establish one level of indirection by making BRL-CAD project
forks of the upstream repositories (or, if the parent project isn't set
up to fork, establish our own as a place to hold that project's files
and updates.)  This gives us a degree of insulation should the upstream
project disappear, but (for those projects were we can properly fork)
also gives us an easy mechanism to pull updates from upstream into our
local forks.

This commit is the first test of the pattern in that it points to a
BRL-CAD/patch fork of https://github.com/shannonbooth/patch - in this
case we simply point at the default git submodule target, since this
project doesn't (yet) have a release we are targeting.