xerces-c needs to be able to find both icu libs
Problem was a little different than I thought - xerces-c wasn't
correctly linking to the icu libraries because libicuuc needs its data
library, and without the rpath help it can't find it in the non-standard
location when it comes time to link - problem manifested when xerces-c
was built with its test applications, so not a BRL-CAD specific problem.
Setting LD_LIBRARY_PATH also resolved the issue, making it clear the
problem was finding the needed libraries.

Needs a full clean build test, but setting the install RPATH for
xerces-c seems to work in initial testing.