See if we can turn off ZSTD for our LLVM use case
Yikes.  ZSTD is causing linking failures when trying to build on
Windows, and hunting around online finds a long history of problems with
the CMake build setup for this use case... see, for example:

https://github.com/llvm/llvm-project/issues/58558

Not quite sure yet what properties aren't getting set on the imported
targets to cause NOTFOUND entries to try and get linked.  If we do need
ZSTD I'll have to try and follow the logic flow through the LLVM CMake
and prepare a patch.