Avoid having git status check submodules
For this repository, the files were are maintaining are not the submodules but
the build logic and patch files.  If we've cloned all the submodules the file
hierarchy is rather large and a status check can be needlessly slow - use the
ignore submodule directive to focus the command on the relevant parts of the
repository.