Start a file for setup instructions
The quick start assumes a system with the necessary dependencies in
place, but unlike the old-school BRL-CAD build we can't stand up these
software stacks without a sizable set of development packages in place.
Need to collect the instructions on how to do that somewhere, since we
aren't aiming to bootstrap things like fontconfig/freetype or XCB X11.