For now, disable OSL shaders and tests build
The tests make sense to turn off, but it's less clear long term if we
really want to disable the shader building.

The immediate motivation is that oslc isn't running correctly in the
build on Windows.  So far it looks as if this is because the directory
containing the oslc.exe executable doesn't have all the dlls it needs -
if we copy oslc.exe and its dlls into the bin directory for
bext_output/install we do get as far as printing the help message (which
does NOT happen with its current location.)  We can probably get past
this by patching the OSL build to copy the dll files from bext_output
into the right place up front, but not bothering with that right now
since it's not 100% sure yet that we need the shaders build for our
purposes.