Start working on updating to 8.x openNURBS branch
Rhino 8 is out, so we need to look at making the jump to the 8.x branch.
This is a first cut at updating the patch files.  Note that the 8.x
branch has CMake logic already in place, so we need to look at minimally
modding what is there rather than overwriting to use our original
CMakeLists.txt file (which is what opennurbs_cmake.patch is currently
doing.)

The lock, lookup and statics patches are disabled, but currently left in
place until we're sure we don't need some variant of them in the new
code.