Get latest appleseed changes
This combination of LLVM, OSL and Appleseed compiles on Linux, and will
build the art binary on Linux.  However, art output is incorrect - it
produces only a black silhouette with moss.g  Going ahead and
committing progress thus far since we seem to be relatively close to
getting it working.