Fix linux build failure affecting tcl/tk, make jobserver related.
Build was failing during Tcl install phase, which was a plain
"make install" and resulting in an error:

make[3]: *** read jobs pipe: Bad file descriptor.  Stop.
make[3]: *** Waiting for unfinished jobs....

This error is related to the make jobserver feature, pointing at
a likely cmake bug in the generated make logic, or a bug in make
itself, or an OS named pipe bug.  Bug encountered on VMs as well
as OpenSuse Linux (which may or may not have been a VM).

Unexpected workaround was to actually add the -j flag to the make
calls which disables jobserver mode on recursive make invocations.
Took that opportunity to enforce -j1 on install calls to avoid
any possibility of output race ordering on what should be an
I/O-bound build stage.