Stash some notes and tests on update checking
We ideally want to trigger configure updates, patching repeats, etc.
when git repository updates are pulled down, but because most of our
builds make src copies and apply patches an update into the src git
tree doesn't automatically trigger rebuilding.

Investigating whether a custom update step can be made to do the steps
for us automatically.  The addition of patching makes what needs to
happen for this a bit involved - what's being committed here is just the
beginnings of thinking about how to do what's needed, and is not
functional.