Try patches to avoid problems loading Dll on Windows
The JavaCheck.cmake test was failing on Windows.  Starting from the
known working 7.38.2 BRL-CAD version, worked through the changes and it
looks as if these were the "poison pills."

Not 100% sure of the reason for the failure, but it looks like we may be
able to disable the code in question.  Post effects don't seem to be
used by the code paths we're interested in.  The 3dm attribute change is
a little riskier and needs testing.