Point Manifold at our local PUGIXML if enabled
Was getting a linking failure with pugixml - -lpugixml doesn't work out of the
box if we're using the bundled pugixml build.

Making a patch file for this - not worth pushing upstream, since the proper fix
in this particular case is to teach Manifold to read and write it's own
enhanced glTF files with topology info using the EXT_mesh_manifold extension.
If we can do that we should be able to completely remove the assimp dependency
from Manifold.