get default mac build working with libpng + system libz.
without enable_all=on, png was getting mixed signals as it wasn't defining the prefix (as there was a system libz), but then not actually adding the system libz to the link line.  result was unresolved zlib symbols.  this tell png to turn off libz building, so it'll use the system libz, if our parent config has building zlib turned off.