Since bext is not expected to be system-installed outside of a BRL-CAD build (and doesn't appear to be structured for it right now also), default CMAKE_INSTALL_PREFIX to the build directory (wherein it will create an install and noinstall subdirs).
This addresses an issue where default bext builds with different
configurations (e.g., Release and Debug) won't look by default into a
common path (HOME/bext_output) and find other configuration
products. This was problematic for non-enable-all builds and even
existing partial builds since it would find and prefer to use the
found version in install even with enable-all.  The non-deterministic
nature made for very difficult to reproduce and debug failure states
as well.

Another issue was the default surprise of having files written outside
of the build dir, unspecified by the user.  This was exacerbated on a
system where build dir and home dir existed on different partitions
with different capacity and performance characteristics.