Use file copying to add whole files for ITcl/Itk
Instead of bloating the patch files, per suggestion from Sean stash copies of
files we are adding wholesale and use copy commands to set them up in the
source tree.

Still need to convert the tcl_hdrs patch for Itcl - probably want to make use
of cmake -E copy_directory_if_different for that one - but this should
establish the general pattern.