Warn if REAL_PATH doesn't match PROJECT_SOURCE_DIR
If we're building in a symlinked directory for our working directory, we end up
with mismatched paths being used for outputs if the build logic is calling
realpath for anything.  That in turn means subsequent strclear logic won't be
looking for the right paths when it is trying to do scrubbing.  Not failing
the build outright, but being noisy since this is very likely to break things.