Update manifold to version 3.2.1 + patches through e047ba6
This version uses linalg rather than glm, eliminating a dependency, and also
features improved determinism when it comes to output generation.

With the new debugging output features in this version, we no longer need the
assimp meshIO support or our patch file.

We explicitly disable some Manifold features in our configure for proper
behavior in our build, as well as bumping Clipper to v1.5.2 to match Manifold.
We also need to define CMAKE_WINDOWS_EXPORT_ALL_SYMBOLS for Clipper2 to get a
.lib file expected by the Manifold build.

Note that changes to the internal math in Manifold v3 (in particularl switching
to double from float) mean that certain automatic "simplification" behaviors we
were seeing in v2 manifold now need to be manually handled on the BRL-CAD side
to avoid major performance hits.