Use git submodule SHA1 values to reset build directories on configure
If a submodule has been updated, its build directory must be considered out of
date and in need of recompilation.  Previously this was done manually, but if
the cmake configure process is re-run we can check a stored SHA1 and handle the
reset automatically.

This is still not perfect - if a rebuild has other implications for the overall
build (like, say, new header versions being used in the updated module that
have compile time implications for other submoduiles using the output of the
changed module) this won't automatically trigger redos on all those builds.
Still, it's a step in the right direction and will handle many cases.