Experiment with using git reset and clean to avoid src copying
Generally in BRL-CAD we treat the source tree as unmodifiable by the build
system.  Applying that principle to bext, that means any time we must patch
a source tree (which is more often than not) we make a copy of the original
sources in the build directory and work on that copy.

This works, but has a considerable downside from a diskspace consumption
perspective (not to mention the cost of all the file I/O.)  A possible
alternative is to use the fact that all the target directories are
submodules and take advantage of git's abilities to reset the source
directory to a clean state.  We can combine --reset hard and -clean -fdx
to return a source tree modified by patching to its pristine state,
ensuring a suitable target for patching even if a previous build had
already applied the patches once.

Commiting this just on zlib now as a test and to work through the potential
tradeoffs of this approach.  Doing it this way will eliminate the ability
of devs to modify and tweak the sources in-situ for quick and easy testing, but
there are ways around that. What I'm currently leaning towards is recommending
to developers that if they want to iteratively work with a particular
dependency to develop a patch, they do so by removing the submodule git
repository and temporarily replacing it with their own (WITHOUT committing the
change from a submodule to a local repo.) Then, having their own repository to
work in, they can commit any intermediate testing states they may want to try
as a foundation for the patch-build cycle to the temporary repo.  Since they
will be committed the testing changes will be the target for the reset and
clean commands and survive multiple patch/build cycles.  They can also do the
reset and clean manually themselves to restore the pre-patch state to make and
commit more changes from a clean, pre-patched state.

Another possible approach would be to see if the patch -R capability could be
utilized at the end of the ExternalProject_Add cycle to undo the patch
applications while still leaving behind any custom mods a developer might have
added while testing.  I'm leaning against trying that because it adds
complexity and fragility to the process - for example, if something went wrong
mid-build with the external project we'd never get to the patch -R step to do
the cleanup and the next cycle through patching would fail.  Plus we could be
asking quite a lot of the patch -R capability depending on what pre-patch
changes were present.  The git reset+clean should be a simple and robust way to
ensure a clean starting point, with the substitute-local-repo-for-submodule
trick allowing for more flexibility on the (usually fairly rare) occasions we
need to do iterative work on a dependency in the bext context.